requirements.txt
llama-cpp-python
Pyyaml
requests